Extract duplicate test helpers into a shared scope
